requirements.txt
Django==5.2
asgiref==3.8.1
django-ckeditor==6.7.2
django-js-asset==3.1.2
django-taggit==6.1.0
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1

